requirements.txt
alembic==1.8.1
attrs==21.4.0
bcrypt==3.2.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
cryptography==37.0.4
distro==1.7.0
dnspython==2.2.1
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2

email-validator==1.2.1
Flask==2.1.3
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==3.2.0
Mako==1.2.1
MarkupSafe==2.1.1
mysqlclient==2.1.1
paramiko==2.11.0
pycparser==2.21
PyMySQL==1.0.2
PyNaCl==1.5.0
pyrsistent==0.18.1
python-dotenv==0.20.0
PyYAML==5.4.1
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.39
texttable==1.6.4
urllib3==1.26.10
websocket-client==0.59.0
Werkzeug==2.1.2
WTForms==3.0.1
